# Conflicts:
#	browser/app/profile/firefox.js
#	browser/branding/aurora/content/firefox-wordmark.svg
#	browser/branding/nightly/content/firefox-wordmark.svg
#	browser/branding/official/content/firefox-wordmark.svg
#	browser/branding/unofficial/PrivateBrowsing_150.png
#	browser/branding/unofficial/PrivateBrowsing_70.png
#	browser/branding/unofficial/content/firefox-wordmark.svg
#	browser/branding/unofficial/pbmode.ico
#	browser/components/about/AboutRedirector.cpp
#	browser/components/preferences/config/permissions-data.mjs
#	browser/components/preferences/config/search.mjs
#	browser/components/preferences/home.inc.xhtml
#	browser/extensions/newtab/content-src/components/Base/Base.jsx
#	browser/extensions/newtab/lib/AboutPreferences.sys.mjs
#	browser/locales/en-US/browser/migrationWizard.ftl
#	browser/locales/en-US/browser/newtab/newtab.ftl
#	browser/locales/en-US/browser/preferences/preferences.ftl
#	browser/locales/en-US/browser/spotlight.ftl
#	devtools/client/debugger/src/components/PrimaryPanes/SourcesTree.js
#	devtools/client/debugger/src/components/PrimaryPanes/SourcesTreeItem.js
#	docshell/base/nsDocShellLoadState.h
#	dom/script/ScriptLoader.cpp
#	python/mozlint/mozlint/types.py
#	remote/webdriver-bidi/test/xpcshell/xpcshell.toml
#	toolkit/locales/en-US/toolkit/branding/brandings.ftl

+++ gfx/cairo/cairo/src/cairo-truetype-subset.c
@@ -114,7 +114,7 @@ check (tt_glyph_data_t,	26);
 
 static cairo_status_t
 cairo_truetype_font_use_glyph (cairo_truetype_font_t	    *font,
-	                       unsigned short		     glyph,
+	                       unsigned long		     glyph,
 			       unsigned short		    *out);
 
 #define SFNT_VERSION			0x00010000
@@ -1026,7 +1026,7 @@ cairo_truetype_font_generate (cairo_truetype_font_t  *font,
 
 static cairo_status_t
 cairo_truetype_font_use_glyph (cairo_truetype_font_t	    *font,
-	                       unsigned short		     glyph,
+	                       unsigned long		     glyph,
 			       unsigned short		    *out)
 {
     if (glyph >= font->base.num_glyphs_in_face)
@@ -1150,7 +1150,7 @@ cairo_truetype_subset_init_internal (cairo_truetype_subset_t     *truetype_subse
 	return status;
 
     for (i = 0; i < font->scaled_font_subset->num_glyphs; i++) {
-	unsigned short parent_glyph = font->scaled_font_subset->glyphs[i];
+	unsigned long parent_glyph = font->scaled_font_subset->glyphs[i];
 	status = cairo_truetype_font_use_glyph (font, parent_glyph, &parent_glyph);
 	if (unlikely (status))
 	    goto fail1;
