+++ gfx/wr/webrender/src/picture_graph.rs
@@ -5,7 +5,8 @@
 use crate::frame_builder::FrameBuildingContext;
 use crate::internal_types::FastHashMap;
 use crate::prim_store::PictureIndex;
-use crate::picture::{PictureInstance, SurfaceIndex, SurfaceInfo};
+use crate::picture::PictureInstance;
+use crate::surface::{SurfaceIndex, SurfaceInfo};
 use crate::tile_cache::{TileCacheInstance, SliceId};
 use smallvec::SmallVec;
 
@@ -83,7 +84,7 @@ impl PictureGraph {
         for (pic_index, info) in self.pic_info.iter().enumerate() {
             if let Some(update_pass) = info.update_pass {
                 let pass = &mut self.update_passes[update_pass];
-                pass.push(PictureIndex(pic_index));
+                pass.push(PictureIndex(pic_index as u32));
             }
         }
     }
@@ -98,17 +99,17 @@ impl PictureGraph {
     ) {
         for pass in &self.update_passes {
             for pic_index in pass {
-                let parent = self.pic_info[pic_index.0].parent;
+                let parent = self.pic_info[pic_index.0 as usize].parent;
 
                 let parent_surface_index = parent.map(|parent| {
                     // Can unwrap here as by the time we have a parent that parent's
                     // surface must have been assigned.
-                    self.pic_info[parent.0].surface_index.unwrap()
+                    self.pic_info[parent.0 as usize].surface_index.unwrap()
                 });
 
-                let info = &mut self.pic_info[pic_index.0];
+                let info = &mut self.pic_info[pic_index.0 as usize];
 
-                match pictures[pic_index.0].assign_surface(
+                match pictures[pic_index.0 as usize].assign_surface(
                     frame_context,
                     parent_surface_index,
                     tile_caches,
@@ -134,19 +135,19 @@ impl PictureGraph {
     ) {
         for pass in self.update_passes.iter().rev() {
             for pic_index in pass {
-                let parent = self.pic_info[pic_index.0].parent;
+                let parent = self.pic_info[pic_index.0 as usize].parent;
 
-                let surface_index = self.pic_info[pic_index.0]
+                let surface_index = self.pic_info[pic_index.0 as usize]
                     .surface_index
                     .expect("bug: no surface assigned during propagate_bounding_rects");
 
                 let parent_surface_index = parent.map(|parent| {
                     // Can unwrap here as by the time we have a parent that parent's
                     // surface must have been assigned.
-                    self.pic_info[parent.0].surface_index.unwrap()
+                    self.pic_info[parent.0 as usize].surface_index.unwrap()
                 });
 
-                pictures[pic_index.0].propagate_bounding_rect(
+                pictures[pic_index.0 as usize].propagate_bounding_rect(
                     surface_index,
                     parent_surface_index,
                     surfaces,
@@ -168,8 +169,8 @@ fn assign_update_pass(
     max_pass_index: &mut usize,
     frame_context: &FrameBuildingContext
 ) {
-    let pic = &mut pictures[pic_index.0];
-    let info = &mut pic_info[pic_index.0];
+    let pic = &mut pictures[pic_index.0 as usize];
+    let info = &mut pic_info[pic_index.0 as usize];
 
     info.parent = parent_pic_index;
