+++ security/manager/ssl/qwac_trust_anchors/src/trust_anchors.rs
@@ -4,7 +4,7 @@
 
 pub(crate) struct TrustAnchor {
     bytes: &'static [u8],
-    subject: (u16, u8),
+    subject: (u16, u16),
 }
 
 impl TrustAnchor {
